Restoring Phase1 changes
